% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctv_suggestions.R
\name{ctv_suggestions}
\alias{ctv_suggestions}
\title{Output suggestions for packages to be added to a Task View}
\usage{
ctv_suggestions(taskview = "Econometrics", n = 5)
}
\arguments{
\item{taskview}{A character vector with one element, must be one of the \href{https://github.com/cran-task-views/ctv#available-task-views}{Task Views available}}

\item{n}{An integer that decides the number of suggestions to show.}
}
\value{
A data frame with suggested packages and there classification probability
}
\description{
Output suggestions for packages to be added to a Task View
}
