% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTVsuggestTrain-package.R
\docType{package}
\name{CTVsuggestTrain-package}
\alias{CTVsuggestTrain}
\alias{CTVsuggestTrain-package}
\title{CTVsuggestTrain: Builds Recommendation Model for CTVsuggest Package}
\description{
Contains functions to construct a model which aims to classify packages into Task Views. The predictions of the model are outputted by the CTVsuggest package.
}
\author{
\strong{Maintainer}: Dylan Dijk \email{zl2291@bristol.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
