% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CRAN_logs.R
\name{get_CRAN_logs}
\alias{get_CRAN_logs}
\title{Title}
\usage{
get_CRAN_logs(TEST = FALSE, limiting_n_observations = 100)
}
\arguments{
\item{limiting_n_observations}{}
}
\value{

}
\description{
Title
}
