% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NLP.R
\name{get_NLP}
\alias{get_NLP}
\title{Creates the NLP features for the model using Task View text and Package description text}
\usage{
get_NLP(TEST = FALSE, limiting_n_observations = 100)
}
\arguments{
\item{TEST}{Takes either TRUE or FALSE, if FALSE it use all available CRAN data otherwise it is restricted to a subset.}
}
\description{
Reads the Task View text from the markdown files on GitHub then cleans it. Then with the corpus of the Task Views, it creates the TF_IDF vectors for each of the words.
}
