% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_model.R
\name{Train_model}
\alias{Train_model}
\title{Title}
\usage{
Train_model(
  TEST = FALSE,
  limiting_n_observations = 100,
  get_input_stored = FALSE,
  get_input_path =
    "tests/testthat/fixtures/get_CRAN_logs_output/get_CRAN_logs_output.rds",
  save = FALSE,
  path = "OUTPUT/"
)
}
\arguments{
\item{save}{logical. If TRUE then saves objects that are returned by this function to disk.}
}
\description{
Title
}
