% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_taskview_data.R
\name{download_taskview_data}
\alias{download_taskview_data}
\title{Title}
\usage{
download_taskview_data(repos = getOption("repos")[1])
}
\arguments{
\item{repos}{}
}
\value{
Data of the current Task Views available
}
\description{
Title
}
\examples{
\donttest{
tvdb = download_taskview_data()
}
}
