% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CRAN_logs.R
\name{get_CRAN_logs}
\alias{get_CRAN_logs}
\title{Title}
\usage{
get_CRAN_logs(
  TEST = FALSE,
  limiting_n_observations = 100,
  get_input_stored = FALSE,
  get_input_path =
    "tests/testthat/fixtures/get_create_features_output/get_create_features_output.rds",
  save_output = FALSE,
  save_path = "tests/testthat/fixtures/get_CRAN_logs_output",
  file_name = "get_CRAN_logs_output.rds"
)
}
\arguments{
\item{limiting_n_observations}{}
}
\description{
Title
}
