% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_create_features.R
\name{get_create_features}
\alias{get_create_features}
\title{Title}
\usage{
get_create_features(
  TEST = FALSE,
  limiting_n_observations = 100,
  get_input_stored = FALSE,
  get_input_path = "tests/testthat/fixtures/get_NLP_output/get_NLP_output.rds",
  save_output = FALSE,
  save_path = "tests/testthat/fixtures/get_create_features_output",
  file_name = "get_create_features_output.rds"
)
}
\arguments{
\item{TEST}{Takes either TRUE or FALSE, if FALSE it use all available CRAN data otherwise it is restricted to a subset.}

\item{limiting_n_observations}{Integer that decides the size of the subset of \code{CRAN_data}, when \code{TEST} is \code{\link{TRUE}}.}
}
\description{
Title
}
