% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_create_features.R
\name{get_create_features}
\alias{get_create_features}
\title{Creates a set of features for each package, using the data that has been extracted from CRAN.}
\usage{
get_create_features(
  TEST = FALSE,
  limiting_n_observations = 100,
  get_input_stored = FALSE,
  get_input_path = "tests/testthat/fixtures/get_NLP_output/get_NLP_output.rds",
  save_output = FALSE,
  save_path = "tests/testthat/fixtures/get_create_features_output",
  file_name = "get_create_features_output.rds"
)
}
\arguments{
\item{TEST}{logical. Default is \code{\link{FALSE}}. If \code{\link{TRUE}}, then a subset of the data that is extracted from CRAN is selected. This is to speed up testing.

More precisely, if \code{\link{TRUE}} a random selection of rows from \code{CRAN_data} is selected, where the number of rows
is given by \code{limiting_n_observations}.}

\item{limiting_n_observations}{Integer that decides the size of the subset of \code{CRAN_data}, when \code{TEST} is \code{\link{TRUE}}.}

\item{get_input_stored}{logical. If \code{\link{TRUE}} then the function uses pre saved data as input, otherwise it runs the \code{CTVsuggestTrain }internal \code{\link[=get_NLP]{get_NLP()}} function.}

\item{get_input_path}{string. If \code{get_input_stored} is set to \code{\link{TRUE}}, \code{get_input_path} gives the path location of the pre saved data.}

\item{save_output}{logical. Default is \code{\link{FALSE}}. If \code{\link{TRUE}}, then the list that is returned is saved to the path set by
\code{save_path}.}

\item{save_path}{string. Sets the path where the list created by the function will be saved,
which is when \code{save_output} is set to \code{\link{TRUE}}}

\item{file_name}{string. Sets the file name for the saved object.}
}
\value{
Returns
\itemize{
\item response_matrix - Matrix with a row for each CRAN package, and a column for each CRAN Task View.
A value of 1 denotes that the package is assigned to the Task View of the corresponding column, and a value of zero if not.
\item features - Matrix with a row for each CRAN package, and a column for each variable
\item All_data - List containing a package and an author network created with cranly.
\item pac_network_igraph - igraph version of the cranly package network.
\item input_CRAN_data - This just a list containing all of the data created by the CTVsuggest:::get_NLP function, so that it is carried forward.
}
}
\description{
In particular this function returns a set of features, and a response matrix for all packages whose data has been extracted.
A subset of these are then used for model training and testing. And the features of the packages are required to generate model predictions, using the trained model.
}
