% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_create_features.R
\name{get_create_features}
\alias{get_create_features}
\title{Title}
\usage{
get_create_features(TEST = FALSE, limiting_n_observations = 100)
}
\arguments{
\item{limiting_n_observations}{}
}
\description{
Title
}
