% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NLP.R
\name{get_NLP}
\alias{get_NLP}
\title{Creates the NLP features for the model using Task View text and Package description text}
\usage{
get_NLP(TEST = FALSE, limiting_n_observations = 100)
}
\arguments{
\item{TEST}{Takes either TRUE or FALSE, if FALSE it use all available CRAN data otherwise it is restricted to a subset.}

\item{limiting_n_observations}{Integer that decides the size of the subset of \code{CRAN_data}, when \code{TEST} is \code{\link{TRUE}}.}
}
\value{
Returns
\itemize{
\item feature_matrix_titles_descriptions_packages_cosine - list. With element for each package being a vector of length of the number of task views.
Each element of the vector is generated by taking the cosine similarity of the TF_IDF vector of the corresponding Task View with the package text data TF_IDF
vector. The IDF term for the TF_IDF vector of the package text is generated by the Task View text corpus.
\item input_CRAN_data - This just a list containing all of the data created by the CTVsuggest:::get_data function, so that it is carried forward.
}
}
\description{
Reads the Task View text from the source markdown files on GitHub, then cleans them.
Then with the corpus of words generated from the Task Views text, it creates TF_IDF vectors for each of the words.
}
