% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_NLP.R
\name{get_NLP}
\alias{get_NLP}
\title{Title}
\usage{
get_NLP(TEST = FALSE, limiting_n_observations = 100)
}
\arguments{
\item{TEST}{Takes either TRUE or FALSE, if FALSE it use all available CRAN data otherwise it is restricted to a subset.}
}
\description{
Title
}
