% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_or_return_objects}
\alias{save_or_return_objects}
\title{Helper function that returns or saves objects created by functions in this package.}
\usage{
save_or_return_objects(
  TEST,
  list_to_return,
  limiting_n_observations,
  save_output,
  save_path,
  file_name
)
}
\arguments{
\item{list_to_return}{}
}
\description{
Helper function that returns or saves objects created by functions in this package.
}
