% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_model.R
\name{Train_model}
\alias{Train_model}
\title{Trains model}
\usage{
Train_model(
  TEST = FALSE,
  limiting_n_observations = 100,
  get_input_stored = FALSE,
  get_input_path =
    "tests/testthat/fixtures/get_CRAN_logs_output/get_CRAN_logs_output.rds",
  save = FALSE,
  path = "OUTPUT/"
)
}
\arguments{
\item{TEST}{Logical. If FALSE model is trained using all available CRAN data, otherwise it is restricted to a subset.}

\item{limiting_n_observations}{Integer. Decides the size of the subset of \code{CRAN_data}, when \code{TEST} is \code{\link{TRUE}}.}

\item{get_input_stored}{Integer. Decides the size of the subset of \code{CRAN_data}, when \code{TEST} is \code{\link{TRUE}}.}

\item{get_input_path}{Integer. Decides the size of the subset of \code{CRAN_data}, when \code{TEST} is \code{\link{TRUE}}.}

\item{save}{Logical. If \code{\link{TRUE}} then saves objects that are returned by this function to disk.}

\item{path}{Logical. If \code{\link{TRUE}} then saves objects that are returned by this function to disk.}
}
\value{
Returns
\itemize{
\item model - Model object
\item predicted_probs_for_suggestions -
\item model accuracy - A percentage value which says how accurate the model is on a test set.
}
}
\description{
Trains model
}
\examples{
\dontrun{
Train_model()
}
}
