% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Train_model.R
\name{Train_model}
\alias{Train_model}
\title{Title}
\usage{
Train_model(TEST = FALSE, limiting_n_observations = 100, save = FALSE)
}
\arguments{
\item{save}{logical. If TRUE then saves objects that are returned by this function to disk.}
}
\value{

}
\description{
Title
}
